extracted Option component
